fix: Elastic error on delete private repo

+++ client/src/containers/RepositoryOptions/index.jsx
@@ -77,7 +77,7 @@ class RepositoryOptions extends React.Component {
   onClickDelete = () => {
     const {
       owner,
-      repoInfo: { id }
+      repoInfo: { id, isPublic }
     } = this.state;
     const { deleteRepo, history } = this.props;
     const { oldName } = this;
@@ -86,7 +86,9 @@ class RepositoryOptions extends React.Component {
       owner,
       name: oldName
     });
-    elasticHelper.deleteRepo(id);
+    if (isPublic) {
+      elasticHelper.deleteRepo(id);
+    }
     history.push(`/${owner}`);
   };
